pyproject.toml
[tool.poetry]
name = "intern"
version = "0.1.0"
description = "my_app"
authors = ["Ser-D <ipc.avanti@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
fastapi-limiter = "^0.1.6"
redis = "^5.0.7"
alembic = "^1.13.2"
aioredis = "^2.0.1"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pytest = "^8.2.2"
pydantic-extra-types = "^2.9.0"
phonenumbers = "^8.13.40"
fastapi-pagination = "^0.12.26"
python-jose = "^3.3.0"
requests = "^2.32.3"
cryptography = "^42.0.8"
apscheduler = "^3.10.4"
pytest-asyncio = "^0.23.8"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"
faker = "^26.3.0"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
pre-commit = "^3.7.1"
click = "^8.1.7"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
